+++ src/app/shell/personal-cabinet/provider/create-workshop/create-about-form/create-about-form.component.ts
@@ -174,7 +174,7 @@ export class CreateAboutFormComponent implements OnInit, OnDestroy {
       this.availableSeatsRadioBtnControl.setValue(false);
     }
 
-    if (this.route.snapshot.paramMap.get('entity') === 'workshop') {
+    if (!this.route.snapshot.paramMap.get('entity')) {
       this.listenToChanges();
     }
