+++ src/app/shell/personal-cabinet/provider/provider-drafts/provider-drafts.component.ts
@@ -3,11 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Store } from '@ngxs/store';
 import { MatDialog } from '@angular/material/dialog';
 import { MatTabChangeEvent } from '@angular/material/tabs';
-import { Constants, ModeConstants, PaginationConstants } from 'shared/constants/constants';
 import { NavBarName } from 'shared/enum/enumUA/navigation-bar';
-import { PaginationElement } from 'shared/models/pagination-element.model';
-import { SearchResponse } from 'shared/models/search.model';
-import { WorkshopCardParameters, WorkshopDraftCard } from 'shared/models/workshop.model';
 import { PushNavPath } from 'shared/store/navigation.actions';
 import {
   DeleteWorkshopDraftById,
@@ -26,16 +22,6 @@ import { ProviderComponent } from '../provider.component';
   styleUrls: ['./provider-drafts.component.scss']
 })
 export class ProviderDraftsComponent extends ProviderComponent implements OnInit, OnDestroy {
-  public readonly constants: typeof Constants = Constants;
-  public readonly ModeConstants = ModeConstants;
-
-  public workshopDrafts: SearchResponse<WorkshopDraftCard[]>;
-  public currentPage: PaginationElement = PaginationConstants.firstPage;
-  public workshopCardParameters: WorkshopCardParameters = {
-    providerId: '',
-    size: PaginationConstants.WORKSHOPS_PER_PAGE
-  };
-
   constructor(
     protected store: Store,
     protected matDialog: MatDialog,
